+++ nginx/Dockerfile
@@ -1,0 +1,1 @@
+FROM nginx:1.27.1
