+++ nginx/conf/monitor_docker.conf.example
