+++ sparky/src/connectors.py
