+++ sparky/src/mongo_connectors.py
