+++ sparky/src/schemas.py
